fix: move bindInput function to buttons.cpp
